# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Signalcraft Coop API
  version: 0.4.0
  description: API for custom orders, store orders, and design review workflows.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://staging.api.signalcraftcoop.com
    description: Staging
  - url: https://api.signalcraftcoop.com
    description: Production
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Orders
  - name: Designs
  - name: Products
  - name: Checkout
  - name: Admin
  - name: Contact
paths:
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /auth/identity-pool-config:
    get:
      tags: [Auth]
      summary: Get Cognito identity pool configuration
      security: []
      responses:
        '200':
          description: Identity pool config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityPoolConfig'
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/OrderType'
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
    post:
      tags: [Orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order details
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /orders/{orderId}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatus'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /orders/{orderId}/designs:
    get:
      tags: [Designs]
      summary: List designs for an order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Designs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Design'
        '401':
          description: Unauthorized
        '404':
          description: Not found
    post:
      tags: [Designs]
      summary: Create design version
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesign'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Design'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /designs/{designId}/reviews:
    post:
      tags: [Designs]
      summary: Submit design review
      parameters:
        - in: path
          name: designId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignReview'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DesignReview'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /products:
    get:
      tags: [Products]
      summary: List active products
      security: []
      responses:
        '200':
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /checkout:
    post:
      tags: [Checkout]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckout'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSession'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
  /contact:
    post:
      tags: [Contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactMessage'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactMessage'
        '400':
          description: Invalid request
  /admin/orders:
    get:
      tags: [Admin]
      summary: List orders (admin)
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/OrderType'
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/orders/{orderId}/status:
    patch:
      tags: [Admin]
      summary: Update order status (admin)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatus'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/orders/{orderId}/shipping:
    patch:
      tags: [Admin]
      summary: Update order shipping info (admin)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderShipping'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/orders/{orderId}/events:
    post:
      tags: [Admin]
      summary: Add order event (admin)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderEvent'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEvent'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/orders/{orderId}/designs:
    get:
      tags: [Admin]
      summary: List designs for an order (admin)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Designs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Design'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    post:
      tags: [Admin]
      summary: Create design version (admin)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesign'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Design'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/messages:
    get:
      tags: [Admin]
      summary: List contact messages (admin)
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactMessage'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/messages/{messageId}:
    patch:
      tags: [Admin]
      summary: Update contact message (admin)
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactMessage'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactMessage'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/products:
    get:
      tags: [Admin]
      summary: List all products (admin)
      responses:
        '200':
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags: [Admin]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/products/{productId}:
    patch:
      tags: [Admin]
      summary: Update product (admin)
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProduct'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/products/{productId}/deactivate:
    patch:
      tags: [Admin]
      summary: Deactivate product (admin)
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OrderType:
      type: string
      enum: [custom, store]
    OrderStatus:
      type: string
      enum: [submitted, designing, review, approved, production, shipping, complete, on-hold, canceled]
    PaymentStatus:
      type: string
      enum: [unpaid, authorized, paid, refunded, disputed]
    DesignStatus:
      type: string
      enum: [draft, in-review, changes-requested, approved]
    DesignReviewStatus:
      type: string
      enum: [approved, changes-requested]
    IdentityPoolConfig:
      type: object
      properties:
        identityPoolId:
          type: string
        region:
          type: string
    User:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        groups:
          type: array
          items:
            type: string
    Address:
      type: object
      required: [line1, city, state, postalCode, country]
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    NFCConfig:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        kind:
          type: string
        notes:
          type: string
    OrderItem:
      type: object
      required: [title, quantity, unitPrice]
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        title:
          type: string
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          minimum: 0
        nfcConfig:
          $ref: '#/components/schemas/NFCConfig'
        metadata:
          type: object
          additionalProperties: true
        designId:
          type: string
          format: uuid
    Order:
      type: object
      required: [id, orderNumber, type, status, paymentStatus, items, subtotal, tax, shipping, total, createdAt]
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        type:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        paymentRequiredAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        paymentProvider:
          type: string
        paymentReference:
          type: string
        paymentMethod:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal:
          type: number
          minimum: 0
        tax:
          type: number
          minimum: 0
        shipping:
          type: number
          minimum: 0
        total:
          type: number
          minimum: 0
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
        shippingCarrier:
          type: string
        shippingService:
          type: string
        trackingNumber:
          type: string
        trackingUrl:
          type: string
          format: uri
        shippedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: '#/components/schemas/OrderEvent'
        createdAt:
          type: string
          format: date-time
    OrderEvent:
      type: object
      required: [id, orderId, type, title, createdAt]
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        description:
          type: string
        metadata:
          type: object
          additionalProperties: true
        isCustomerVisible:
          type: boolean
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    Product:
      type: object
      required: [id, title, sku, basePrice, allowsNfc]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        sku:
          type: string
        description:
          type: string
        basePrice:
          type: number
          minimum: 0
        category:
          type: string
        allowsNfc:
          type: boolean
        allowsLogoUpload:
          type: boolean
        active:
          type: boolean
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        createdAt:
          type: string
          format: date-time
    ProductImage:
      type: object
      required: [url]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        isMain:
          type: boolean
        sortOrder:
          type: integer
        altText:
          type: string
    Design:
      type: object
      required: [id, orderId, version, status, previewUrl, createdAt]
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/DesignStatus'
        previewUrl:
          type: string
          format: uri
        sourceUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    DesignReview:
      type: object
      required: [id, designId, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        designId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/DesignReviewStatus'
        comment:
          type: string
        attachmentUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    ContactMessage:
      type: object
      required: [id, name, email, message, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        message:
          type: string
        status:
          type: string
          enum: [new, in-progress, resolved, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrder:
      type: object
      required: [type, items]
      properties:
        type:
          $ref: '#/components/schemas/OrderType'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
    UpdateOrderStatus:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
    UpdateOrderShipping:
      type: object
      properties:
        shippingCarrier:
          type: string
        shippingService:
          type: string
        trackingNumber:
          type: string
        trackingUrl:
          type: string
          format: uri
        shippedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
    CreateOrderEvent:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        description:
          type: string
        metadata:
          type: object
          additionalProperties: true
        isCustomerVisible:
          type: boolean
    CreateDesign:
      type: object
      required: [version, status, previewUrl]
      properties:
        version:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/DesignStatus'
        previewUrl:
          type: string
          format: uri
        sourceUrl:
          type: string
          format: uri
    CreateDesignReview:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/DesignReviewStatus'
        comment:
          type: string
        attachmentUrl:
          type: string
          format: uri
    CreateContactMessage:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        message:
          type: string
    UpdateContactMessage:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, in-progress, resolved, archived]
    CreateProduct:
      type: object
      required: [title, sku, basePrice, allowsNfc]
      properties:
        title:
          type: string
        sku:
          type: string
        description:
          type: string
        basePrice:
          type: number
          minimum: 0
        category:
          type: string
        allowsNfc:
          type: boolean
        allowsLogoUpload:
          type: boolean
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
    UpdateProduct:
      type: object
      properties:
        title:
          type: string
        sku:
          type: string
        description:
          type: string
        basePrice:
          type: number
          minimum: 0
        category:
          type: string
        allowsNfc:
          type: boolean
        allowsLogoUpload:
          type: boolean
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
    CreateCheckout:
      type: object
      required: [orderId, successUrl, cancelUrl]
      properties:
        orderId:
          type: string
          format: uuid
        successUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
    CheckoutSession:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri